moved player to separate package
